requirements.txt
influxdb_client
python-dotenv
ping3
speedtest-cli